Move editions into their own view file
