+++ bookwyrm/views/__init__.py
@@ -4,11 +4,12 @@ from .authentication import Login, Register, Logout
 from .authentication import ConfirmEmail, ConfirmEmailCode, resend_link
 from .author import Author, EditAuthor
 from .block import Block, unblock
-from .books import Book, EditBook, ConfirmEditBook, Editions
-from .books import upload_cover, add_description, switch_edition, resolve_book
+from .books import Book, EditBook, ConfirmEditBook
+from .books import upload_cover, add_description, resolve_book
 from .directory import Directory
 from .discover import Discover
 from .edit_user import EditUser, DeleteUser
+from .editions import Editions, switch_edition
 from .federation import Federation, FederatedServer
 from .federation import AddFederatedServer, ImportServerBlocklist
 from .federation import block_server, unblock_server
